requirements.txt
accelerate==1.13.0
aiobotocore==3.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aioitertools==0.13.0
aiosignal==1.4.0
altair==6.0.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
appdirs==1.4.4
appnope==0.1.4 ; sys_platform == 'darwin'
asttokens==3.0.1
asyncssh==2.22.0
atpublic==7.0.0
attrs==26.1.0
backoff==2.2.1
billiard==4.2.4
blinker==1.9.0
boto3==1.42.70
botocore==1.42.70
cachetools==7.0.5
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comm==0.2.3
configobj==5.0.9
cryptography==46.0.5
dacite==1.6.0
dagshub==0.6.9
dagshub-annotation-converter==0.1.16
dataclasses-json==0.6.7
datasets==4.8.3
debugpy==1.8.20
decorator==5.2.1
dictdiffer==0.9.0
dill==0.4.1
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==1.1.0
dvc==3.67.0
dvc-data==3.18.3
dvc-http==2.32.0
dvc-objects==5.2.0
dvc-render==1.0.2
dvc-s3==3.3.0
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
evaluate==0.4.6
executing==2.2.1
fastapi==0.135.2
filelock==3.25.2
flatten-dict==0.4.2
flufl-lock==9.0.0
frozenlist==1.8.0
fsspec==2026.2.0
funcy==2.0
gitdb==4.0.12
gitpython==3.1.46
gql==4.0.0
grandalf==0.8
graphql-core==3.2.8
gto==1.9.0
h11==0.16.0
hf-xet==1.4.2 ; platform_machine == 'AMD64' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'x86_64'
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.7.2
hydra-core==1.3.2
idna==3.11
iniconfig==2.3.0
ipykernel==7.2.0
ipython==9.11.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.8
iterative-telemetry==0.0.10
jedi==0.19.2
jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-client==8.8.0
jupyter-core==5.9.1
jupyterlab-widgets==3.0.16
kombu==5.6.2
lxml==6.0.2
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
mypy-extensions==1.1.0
narwhals==2.18.1
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.3
omegaconf==2.3.0
orjson==3.11.7 ; implementation_name == 'cpython'
packaging==26.0
pandas==2.3.3
parso==0.8.6
pathspec==1.0.4
pathvalidate==3.3.1
pexpect==4.9.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
ptyprocess==0.7.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
pure-eval==0.2.3
pyarrow==23.0.1
pycparser==3.0 ; implementation_name != 'PyPy'
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
pydeck==0.9.1
pydot==4.0.1
pygit2==1.19.1
pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.3.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
pywin32==311 ; sys_platform == 'win32'
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
ruamel-yaml==0.19.1
ruff==0.15.8
s3fs==2026.2.0
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
scmrepo==3.6.1
semver==3.0.4
sentry-sdk==2.55.0
seqeval==1.2.2
setuptools==81.0.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
six==1.17.0
smmap==5.0.3
sqltrie==0.11.2
stack-data==0.6.3
starlette==1.0.0
streamlit==1.55.0
sympy==1.14.0
tabulate==0.10.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
tomlkit==0.14.0
torch==2.11.0 ; sys_platform != 'emscripten' and sys_platform != 'linux' and sys_platform != 'win32'
torch==2.11.0+cpu ; sys_platform == 'emscripten' or sys_platform == 'linux' or sys_platform == 'win32'
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.3.0
treelib==1.8.0
typer==0.24.1
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.42.0
vine==5.1.0
voluptuous==0.16.0
wandb==0.25.0
watchdog==6.0.0 ; sys_platform != 'darwin'
wcwidth==0.6.0
widgetsnbextension==4.0.15
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zc-lockfile==4.0
